v 1.0.6
In this release, the application section for data connectors has been
modified to be interactive. The Clients _include ahas been revised, and
the CSS and JS to support the data connector functionality has been
added. The blog post for data modelling has been changed to
data-modeling to not have %20 in the URL. Additional Tweaks to block
HTML for clients section for better rendering of Mobile pages.